+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## UNRELEASED
+* N/A
+
+## 0.1.0-alpha.2.3
 * `expects` / `exposes`: Add `type: :uuid` special case validation
 * [BUGFIX] Allow `hoist_errors` to pass the result through on success (allow access to subactions' exposures)
 * [`Axn::Factory`] Support error_from + rescues